Create DSign.sol
First attempt to create a digital signature using blockchain. It attempts to solve problem with signing document using digital signature which is not unique to the specific document and can be easily forged. Smart contract can be used to solve this issue. 

Each document first has to be hashed to get the hash256 as input for the smart contract DSign. The hash is unique to the document. 
Only the creator of the document can create number of people who need to sign the document, by putting in a passcode for each person and generate the initial key in the blockchain.  
The passcode should then be informed to the person who needs to sign the document. 
The person with the passcode can now generate his signature key by giving a signature code for hashing purpose, thereby no one except this person has the signature key. Not even the contract owner has the signature key. Thus, making it unique and full ownership control. 
The person can then use the signature key to perform signing of document. The timestamp on the blockchain is the proof that the document is verified and signed by the person.